Sublime Build File for Windows
This file can be used to build an mbx file.  This version is set up to run on my office PC compiling files into a tmp directory which is located in the same main directory as the src directory containing the source files.  
e.g.
assume mathbook is installed to "c:/mathbook/" and compile "~/project_dir/src/file.mbx" in "~/project_dir/tmp/"